Ensure the URL includes a section hash when opening an overlay
There are some cases, like clicking the "Read More" button in the first
item in a carousel, or returning to that button after interacting with a
different part of the page, where the document URL might not have a
secion hash in it when entering an overlay. This makes sure that this
hash fragment is set.