1040: Map: Serialize relevant slide attributes to avoid map re-render
Mapbox GL JS usage is billed by map loads. Reducing this as much
as possible avoids billing impact while editing.

Previously, edits to text within the map slides would cause the
map to rerender because it would be a change to the innerBlocks.
Serializing only the attributes we actually need avoids that, by
giving the useEffect a stable string reference instead of an
oft-changing object.